Move transcript editor component out of profile page
